# apps/admin/src/api/generated/schemas/updateChannel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fox Sphere API
 * REST API for the Fox Sphere Twitch bot platform
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateChannelStatus } from './updateChannelStatus';

export interface UpdateChannel {
  status?: UpdateChannelStatus;
  botIsMod?: boolean;
}